# docs/swagger.yaml
basePath: /api
definitions:
  dto.CreateUserRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  dto.LoginUserRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  dto.UpdateUserRequest:
    properties:
      email:
        type: string
      username:
        type: string
    type: object
  errorhandler.ErrorResponse:
    properties:
      message:
        type: string
    type: object
  handlers.HealthCheckResponse:
    properties:
      errors:
        $ref: '#/definitions/handlers.Services'
      services:
        $ref: '#/definitions/handlers.Services'
      status:
        type: string
      timestamp:
        type: string
    type: object
  handlers.Services:
    properties:
      database:
        type: string
      redis:
        type: string
    type: object
  models.User:
    properties:
      created:
        type: string
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      updated:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample API created by Golang.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Golang REST API
  version: "1.0"
paths:
  /health:
    get:
      consumes:
      - application/json
      description: Check the health status of the server and its dependencies
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.HealthCheckResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/handlers.HealthCheckResponse'
      summary: Health check
      tags:
      - health
  /users:
    get:
      consumes:
      - application/json
      description: Get all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errorhandler.ErrorResponse'
      summary: Get users
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing user by providing user ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorhandler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorhandler.ErrorResponse'
      summary: Delete an user
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Update an existing user by providing username and email
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update User Request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorhandler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorhandler.ErrorResponse'
      summary: Update an user
      tags:
      - users
  /users/login:
    post:
      consumes:
      - application/json
      description: Get user token
      parameters:
      - description: Login User Request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.LoginUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorhandler.ErrorResponse'
      summary: Get user token
      tags:
      - users
  /users/me:
    get:
      consumes:
      - application/json
      description: Get user by providing JWT token
      parameters:
      - description: Bearer {token}
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorhandler.ErrorResponse'
      summary: Get user by JWT token
      tags:
      - users
  /users/register:
    post:
      consumes:
      - application/json
      description: Create a new user by providing username, email and password
      parameters:
      - description: Create User Request
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errorhandler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/errorhandler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errorhandler.ErrorResponse'
      summary: Create an user
      tags:
      - users
swagger: "2.0"